More cleanup, move directory for where tests live
