+++ ihmc-perception/src/test/java/us/ihmc/perception/gpuMapping/HeightMapMessageToolsTest.java
@@ -1,12 +1,8 @@
-package us.ihmc.perception.heightmap;
+package us.ihmc.perception.gpuMapping;
 
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import perception_msgs.msg.dds.HeightMapMessage;
-import us.ihmc.perception.gpuMapping.HeightMapData;
-import us.ihmc.perception.gpuMapping.HeightMapMessageTools;
-import us.ihmc.perception.gpuMapping.HeightMapTools;
 
 import static org.junit.jupiter.api.Assertions.*;
