+++ ihmc-perception/src/test/java/us/ihmc/perception/gpuMapping/TerrainMapExtractorTest.java
@@ -1,4 +1,4 @@
-package us.ihmc.perception.gpuHeightMap;
+package us.ihmc.perception.gpuMapping;
 
 import org.bytedeco.opencv.global.opencv_core;
 import org.bytedeco.opencv.opencv_core.GpuMat;
@@ -10,12 +10,6 @@ import us.ihmc.euclid.geometry.Plane3D;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.Vector3D;
 import us.ihmc.log.LogTools;
-import us.ihmc.perception.gpuMapping.HeightMapParameters;
-import us.ihmc.perception.gpuMapping.HeightMapTools;
-import us.ihmc.perception.gpuMapping.SnapResult;
-import us.ihmc.perception.gpuMapping.TerrainMapParameters;
-import us.ihmc.perception.gpuMapping.TerrainMapData;
-import us.ihmc.perception.gpuMapping.TerrainMapExtractor;
 
 import static org.junit.jupiter.api.Assertions.*;
