+++ ihmc-perception/src/test/java/us/ihmc/perception/gpuMapping/TerrainMapMessageToolsTest.java
@@ -1,4 +1,4 @@
-package us.ihmc.perception.gpuHeightMap;
+package us.ihmc.perception.gpuMapping;
 
 import org.bytedeco.opencv.global.opencv_core;
 import org.bytedeco.opencv.opencv_core.Mat;
@@ -7,9 +7,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import perception_msgs.msg.dds.TerrainMapMessage;
 import us.ihmc.euclid.tuple3D.Point3D;
-import us.ihmc.perception.gpuMapping.TerrainMapData;
-import us.ihmc.perception.gpuMapping.TerrainMapMessageTools;
-import us.ihmc.perception.gpuMapping.TerrainMapTools;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
