+++ ihmc-perception/src/test/java/us/ihmc/perception/gpuMapping/worldmodel/ChunkedMapManagerTest.java
@@ -1,6 +1,5 @@
-package us.ihmc.perception.heightmap;
+package us.ihmc.perception.gpuMapping.worldmodel;
 
-import org.apache.log4j.Hierarchy;
 import org.bytedeco.opencv.global.opencv_core;
 import org.bytedeco.opencv.opencv_core.Mat;
 import org.junit.jupiter.api.Test;
